[fsdp] refactor: move unfuse_moe_params to FSDP backend (#7161)
### What does this PR do?

Move unfuse_moe_params from vllm to FSDP backend

+++ tests/checkpoint_engine/test_utils.py
@@ -83,7 +83,9 @@ class MockServerAdapter(BaseRollout):
         for name, weight in self.model.state_dict().items():
             assert name in self.received_weights, f"weight {name} not received"
             received = self.received_weights[name]
-            assert torch.allclose(weight.to(received.device), received), f"weight {name} not equal"
+            assert torch.allclose(weight.to(received.device, dtype=received.dtype), received), (
+                f"weight {name} not equal"
+            )
         self.received_weights.clear()
         print("Check passed, all weights are equal!")
